# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'TechForb Challenge'","description":"'Challenge para Puesto Trainee'","contact":{"name":"Tobias Moreno","email":"tobiasmoreno.tm.21@gmail.com"},"version":"'1.0.0'"},"servers":[{"url":"http://localhost:8080","description":"'Challenge para Puesto Trainee'"}],"paths":{"/api/readings/update/{id}":{"put":{"tags":["reading-controller"],"summary":"Actualizar una lectura existente","operationId":"updateReading","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestUpdateReadingDTO"}}},"required":true},"responses":{"404":{"description":"Lectura no encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"No autorizado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Lectura actualizada con éxito","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseReadingDTO"}}}},"400":{"description":"Solicitud inválida","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}}}},"/api/readings/create":{"post":{"tags":["reading-controller"],"summary":"Crear una nueva lectura","operationId":"createReading","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestReadingDTO"}}},"required":true},"responses":{"201":{"description":"Lectura creada con éxito","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseReadingDTO"}}}},"401":{"description":"No autorizado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Solicitud inválida","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}}}},"/api/auth/register":{"post":{"tags":["auth-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/auth/refresh-token":{"post":{"tags":["auth-controller"],"operationId":"refreshToken","responses":{"200":{"description":"OK"}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/readings/list":{"get":{"tags":["reading-controller"],"summary":"Obtener todas las lecturas del usuario actual","operationId":"getAllReadings","responses":{"401":{"description":"No autorizado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Lista de lecturas obtenida con éxito","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseReadingDTO"}}}},"403":{"description":"No autorizado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}}}},"/api/readings/get/{id}":{"get":{"tags":["reading-controller"],"summary":"Obtener una lectura por ID","operationId":"getReadingById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Lectura no encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"No autorizado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"Lectura encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseReadingDTO"}}}}}}},"/api/readings/{id}":{"delete":{"tags":["reading-controller"],"summary":"Eliminar una lectura por ID","operationId":"deleteReading","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Lectura no encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"401":{"description":"No autorizado","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"204":{"description":"Lectura eliminada con éxito","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseReadingDTO"}}}}}}}},"components":{"schemas":{"RequestUpdateReadingDTO":{"type":"object","properties":{"readingValue":{"type":"number","format":"double"},"timestamp":{"type":"string","format":"date-time"},"sensorId":{"type":"integer","format":"int64"},"alerts":{"type":"array","items":{"$ref":"#/components/schemas/ResponseAlertDTO"}}}},"ResponseAlertDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string"},"message":{"type":"string"},"reading":{"$ref":"#/components/schemas/ResponseReadingDTO"}}},"ResponseReadingDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"readingValue":{"type":"number","format":"double"},"timestamp":{"type":"string","format":"date-time"},"sensor":{"type":"string"},"alerts":{"type":"array","items":{"$ref":"#/components/schemas/ResponseAlertDTO"}}}},"ApiError":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"RequestAlertDTO":{"type":"object","properties":{"type":{"type":"string"},"message":{"type":"string"},"readingId":{"type":"integer","format":"int64"}}},"RequestReadingDTO":{"type":"object","properties":{"readingValue":{"type":"number","format":"double"},"timestamp":{"type":"string","format":"date-time"},"sensorId":{"type":"integer","format":"int64"},"alerts":{"type":"array","items":{"$ref":"#/components/schemas/RequestAlertDTO"}}}},"RegisterRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthResponse":{"type":"object","properties":{"access_token":{"type":"string"},"refresh_token":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}}}}}